pyproject.toml
[project]
name = "sda-boolean-networks"
version = "0.1.0"
description = "SDA 2 - Final Project"
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "boolean-py>=5.0",
    "matplotlib>=3.10.7",
    "networkx>=3.5",
    "numpy>=2.3.4",
    "tqdm>=4.67.1",
]
